rename. explicitly cache dataset. add low bit optim
